cambios en registrar evento

+++ views/listaEventos.php
@@ -6,22 +6,29 @@ $evento=new Evento();
 $lista=$evento->getInfoEventos();
 $pageTitle="Eventos";
 include_once("../views/header_index.php");?>
-	<div class="container-fluid">
-      <div class="row-fluid">
-		<h1>Eventos</h1>
+<div class="container-fluid">
+	<div class="row-fluid">
+		<div class="span6 offset3">
+			<h1>Eventos</h1>
 			<table class="table table-bordered">
-			<tr><th colspan=3>Titulo</th></tr>
-			<?php			
-			foreach ($lista as $obj) {
-				echo "<tr><td>".$obj["titulo"]."</td>";				
-				echo "<td><a class='btn btn-primary' href='verInformacionEvento.php?id=".$obj["_id"]."'><i class='icon-glass icon-white'></i>Ver Informaci&oacute;n</a></td>";
-				echo "<td><a class='btn btn-danger' href='../controllers/eliminarEvento.php?id=".$obj["_id"]."'><i class='icon-trash icon-white'></i>Eliminar evento</a></td></tr>";					
-			}
-			?>
+				<tr>
+					<th colspan=3>Titulo</th>
+				</tr>
+				<?php			
+				foreach ($lista as $obj) {
+					echo "<tr><td>".$obj["titulo"]."</td>";
+					echo "<td><a class='btn btn-primary' href='verInformacionEvento.php?id=".$obj["_id"]."'><i class='icon-glass icon-white'></i> Ver Informaci&oacute;n</a></td>";
+					if($tipo==0){
+						echo "<td><a class='btn btn-danger' href='../controllers/eliminarEvento.php?id=".$obj["_id"]."'><i class='icon-trash icon-white'></i> Eliminar evento</a></td></tr>";
+					}
+				}
+				?>
 			</table>
-			<a class='btn' href='registraEvento.php'><i class='icon-glass'></i>Registrar evento</a>
+			<a class='btn' href='registraEvento.php'><i class='icon-glass'></i>Registrar
+				evento</a>
 		</div>
-		</div>
-		<?php include_once("../views/footer.html");?>
+	</div>
+</div>
+<?php include_once("../views/footer.html");?>
 </body>
 </html>
